pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bayaml"
version = "0.1.5"
description = "Bayaml is a lightweight ML & AI orchestration framework that automates data loading, preprocessing, model training, evaluation, and deployment through a deterministic, privacy-safe Auto Mode with built-in AutoML. Built and maintained by Aditya Sarode."
readme = "README.md"
requires-python = ">=3.9"
license = { file = "LICENSE" }
authors = [{ name = "Aditya Sarode" }]
keywords = ["machine-learning", "data-science", "mlops", "orchestration", "pipeline", "framework", "ai", "bayaml", "BAYAML"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "Intended Audience :: Science/Research",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",
  "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
  "pandas>=1.5",
  "numpy>=1.22",
  "scikit-learn>=1.2",
  "scipy>=1.10",
  "pyyaml>=6.0",
  "matplotlib>=3.6",
  "tabulate>=0.9.0",
  "jinja2>=3.1.0",
]

[project.optional-dependencies]
dev = ["pytest", "ruff", "mypy"]
viz = ["seaborn", "plotly"]
deep = ["tensorflow", "torch"]

[project.urls]
GitHub = "https://github.com/adityassarode"
Repository = "https://github.com/adityassarode/bayaml"
Issues = "https://github.com/adityassarode/bayaml/issues"
Instagram = "https://www.instagram.com/adityassarode"
Buymeacoffee = "https://www.buymeacoffee.com/adityassarode"
Ko-fi = "https://ko-fi.com/adityassarode"


[project.scripts]
bayaml = "bayaml.cli.main:main"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["bayaml*"]

[tool.setuptools.package-data]
bayaml = ["py.typed", "assets/*.yaml"]

[tool.pytest.ini_options]
addopts = "-ra -q"
testpaths = ["tests"]

[tool.coverage.run]
source = ["bayaml"]

[tool.coverage.report]
show_missing = true
skip_covered = true

[tool.black]
line-length = 88
target-version = ["py39"]

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "B", "I"]
ignore = []

[tool.mypy]
python_version = "3.9"
strict = true
ignore_missing_imports = true
